fix: python dependency conflict
