+++ packages/core/src/python/motia_logger.py
@@ -1,6 +1,6 @@
 import time
 from typing import Any, Dict, Optional
-from rpc import RpcSender
+from motia_rpc import RpcSender
 
 class Logger:
     def __init__(self, trace_id: str, flows: list[str], rpc: RpcSender):
