+++ packages/core/src/python/motia_rpc.py
@@ -1,13 +1,7 @@
-import uuid
-import asyncio
-import json
-import sys
-import os
-import platform
-from typing import Any, Dict, Optional, Callable, Union
-from communication_factory import create_communication
-from rpc_communication import RpcCommunication
-from ipc_communication import IpcCommunication
+from typing import Any, Union
+from motia_communication_factory import create_communication
+from motia_rpc_communication import RpcCommunication
+from motia_ipc_communication import IpcCommunication
 
 def serialize_for_json(obj: Any) -> Any:
     """Convert Python objects to JSON-serializable types"""
